feat: Implement create table and update table api for rest catalog. (#97)
* feat: Implement update table/create table api for rest catalog

* Add create table test

* Add some tests for update table

* Add tests for table updates and requirements

* Add some comments

* Add tests for transaction

* Fix format

* Fix linters

* Stop upgrading to uuid 1.6.0

* Fix comments

+++ crates/iceberg/src/spec/sort.rs
@@ -20,6 +20,7 @@
 */
 use serde::{Deserialize, Serialize};
 use std::sync::Arc;
+use typed_builder::TypedBuilder;
 
 use super::transform::Transform;
 
@@ -47,12 +48,12 @@ pub enum NullOrder {
     Last,
 }
 
-#[derive(Debug, Serialize, Deserialize, PartialEq, Eq, Clone)]
+#[derive(Debug, Serialize, Deserialize, PartialEq, Eq, Clone, TypedBuilder)]
 #[serde(rename_all = "kebab-case")]
 /// Entry for every column that is to be sorted
 pub struct SortField {
     /// A source column id from the table’s schema
-    pub source_id: i64,
+    pub source_id: i32,
     /// A transform that is used to produce values to be sorted on from the source column.
     pub transform: Transform,
     /// A sort direction, that can only be either asc or desc
@@ -61,13 +62,14 @@ pub struct SortField {
     pub null_order: NullOrder,
 }
 
-#[derive(Debug, Serialize, Deserialize, PartialEq, Eq, Clone, Builder)]
+#[derive(Debug, Serialize, Deserialize, PartialEq, Eq, Clone, Builder, Default)]
 #[serde(rename_all = "kebab-case")]
 #[builder(setter(prefix = "with"))]
 /// A sort order is defined by a sort order id and a list of sort fields.
 /// The order of the sort fields within the list defines the order in which the sort is applied to the data.
 pub struct SortOrder {
     /// Identifier for SortOrder, order_id `0` is no sort order.
+    #[builder(default)]
     pub order_id: i64,
     /// Details of the sort
     #[builder(setter(each(name = "with_sort_field")), default)]
